requirements.txt
absl-py==2.1.0
alembic==1.13.1
annotated-types==0.6.0
anyio==3.7.1
APScheduler==3.10.4
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.4.0
beautifulsoup4==4.12.2
brotlicffi==1.0.9.2
certifi==2025.10.5
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.2.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.1
cycler==0.11.0
debugpy==1.8.17
decorator==5.2.1
exceptiongroup==1.2.0
executing==2.2.1
fastapi==0.104.1
fastjsonschema==2.20.0
flatbuffers==24.3.25
fonttools==4.60.1
gast==0.6.0
geographiclib==2.1
geopy==2.4.1
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.71.0
h11==0.16.0
h5py==3.14.0
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
ipykernel==7.0.1
ipython==8.37.0
jedi==0.19.2
joblib==1.5.2
jsonschema==4.25.0
jsonschema-specifications==2023.7.1
jupyter_client==8.6.3
jupyter_core==5.9.1
keras==3.11.2
kiwisolver==1.4.8
Mako==1.3.10
Markdown==3.8
markdown-it-py==2.2.0
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.2.1
mdurl==0.1.2
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
ml_dtypes==0.5.1
mpi4py==4.0.3
namex==0.1.0
narwhals==2.9.0
nbformat==5.10.4
nest_asyncio==1.6.0
numpy==1.26.4
opt-einsum==3.3.0
optree==0.14.1
packaging==25.0
pandas==2.1.4
parso==0.8.5
pickleshare==0.7.5
pillow==11.1.0
pip==25.2
platformdirs==4.5.0
plotly==6.3.1
pluggy==1.6.0
prompt_toolkit==3.0.52
protobuf==5.29.3
psutil==7.1.1
psycopg2-binary==2.9.9
pure_eval==0.2.3
pyarrow==14.0.1
pycparser==2.23
pydantic==2.5.0
pydantic_core==2.14.1
Pygments==2.19.1
pyparsing==3.2.0
PyPDF2==3.0.1
PyQt6==6.9.1
PyQt6_sip==13.10.2
PySocks==1.7.1
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
pywin32==311
PyYAML==6.0.1
pyzmq==27.1.0
redis==5.0.1
referencing==0.30.2
requests==2.32.5
rich==13.9.4
rpds-py==0.22.3
scikit-learn==1.7.1
scipy==1.15.3
seaborn==0.13.2
setuptools==78.1.1
sip==6.12.0
six==1.17.0
sniffio==1.3.0
soupsieve==2.8
SQLAlchemy==2.0.23
stack_data==0.6.3
starlette==0.27.0
tensorboard==2.18.0
tensorboard_data_server==0.7.0
tensorflow_cpu==2.18.1
termcolor==2.1.0
threadpoolctl==3.5.0
tomli==2.2.1
tornado==6.5.1
traitlets==5.14.3
typing_extensions==4.15.0
typing-inspection==0.4.0
tzdata==2025.2
tzlocal==5.2
urllib3==2.5.0
uvicorn==0.24.0
wcwidth==0.2.14
Werkzeug==3.1.3
wheel==0.45.1
win-inet-pton==1.1.0
wrapt==1.17.0
xgboost==3.1.1
zipp==3.23.0
lightgbm==4.6.0
